Remove redundant script (generate-results.py)
-Merge generate-results.py and autoprocess.py scripts
-Rename autoprocess.py to process.py

+++ process.py
@@ -7,6 +7,7 @@ import re
 import os
 import glob
 import subprocess
+import sys
 
 import smtplib
 from dotenv import dotenv_values
@@ -64,6 +65,32 @@ def insert_record(curr, filepath):
                  [filepath, formatted_date, path, checksum])
 
 
+def get_runs(paths, ignore_list, runs_list, callback=None):
+    """
+    Gets the list of runs to process
+
+    :param paths: list, paths to all R1 files in the uploads directory
+    :param ignore_list: list, directories that the user would like to avoid processing
+    :param runs_list: list, directories that the user would like to processing, processes everything if none specified
+    :param callback: function, option to print messages to the console
+    :return: list, paths to all files that have not been inserted into the database and the filepaths to all runs
+    """
+    runs = set()
+    ignored = set()
+    for file in paths:
+        path, filename = os.path.split(file)
+        if contains_run(path, ignore_list):
+            ignored.add(path)
+            continue
+        if len(runs_list) == 0 or contains_run(path, runs_list):
+            runs.add(path)
+
+    for path in ignored:
+        callback("Ignoring '{}'".format(path))
+
+    return runs
+
+
 def get_files(curr, paths, ignore_list, check_processed, callback=None):
     """
     Detects if there are any new data files that have been uploaded by comparing the list of files to those that
@@ -83,7 +110,7 @@ def get_files(curr, paths, ignore_list, check_processed, callback=None):
     entered = []
     ignore = []
     for file in paths:
-        if ignore_file(file, ignore_list):
+        if contains_run(file, ignore_list):
             ignore.append(file)
             continue
         path, filename = os.path.split(file)
@@ -118,16 +145,16 @@ def get_files(curr, paths, ignore_list, check_processed, callback=None):
     return unentered, runs
 
 
-def ignore_file(file, ignore):
+def contains_run(run, usr_list):
     """
-    Ignores file if the file is within any of the user specified directories
+    Ignores/Includes run if its in the user specified directories
 
-    :param file: str, path to a file
-    :param ignore: list, directories to ignore
+    :param run: str, path to a run
+    :param usr_list: list, directories to ignore
     :return: bool, True if the directory is in the file path, False otherwise
     """
-    for directory in ignore:
-        if directory in file:
+    for directory in usr_list:
+        if directory in run:
             return True
     return False
 
@@ -239,13 +266,14 @@ def process_files(curr, indir, outdir, paths, binpath="minimap2", cutabin="cutad
             callback("\t {}".format(f), level="DEBUG")
 
 
-def run_scripts(runs, indir, outdir, callback=None):
+def run_scripts(runs, indir, outdir, replace, callback=None):
     """
     Run estimate-freqs.R, make-barplots.R and make-csv.R for each run
 
     :param runs: list, location of the run files for the estimate-freqs.R script
     :param indir: str, location of the uploads folder
     :param outdir: str, location of the results folder
+    :param replace: bool, True if old files need to be replaced, False otherwise
     :param callback: function, option to print messages to the console
     :return: None
     """
@@ -265,7 +293,7 @@ def run_scripts(runs, indir, outdir, callback=None):
                     sys.exit()
 
             filename = "{}-{}-{}".format(os.path.basename(os.path.dirname(result_dir)),
-                                              os.path.basename(result_dir), lineage.replace('.',''))
+                                         os.path.basename(result_dir), lineage.replace('.',''))
             cmd = ['Rscript', 'scripts/estimate-freqs.R', result_dir, constellation, '{}.json'.format(filename)]
 
             # Get metadata filename, not all files are named "metadata.csv"
@@ -314,9 +342,10 @@ def run_scripts(runs, indir, outdir, callback=None):
                 continue
 
             # Remove prior output files
-            resfiles = glob.glob("{}/**/{}.*".format(result_dir, filename), recursive=True)
-            for resfile in resfiles:
-                os.remove(resfile)
+            if replace:
+                resfiles = glob.glob("{}/**/{}.*".format(result_dir, filename), recursive=True)
+                for resfile in resfiles:
+                    os.remove(resfile)
 
             for suffix in suffixes:
                 stdout = subprocess.getoutput("sha1sum {}.{}".format(filename, suffix))
@@ -346,11 +375,15 @@ def parse_args():
                         help="<option> path to cutadapt executable")
     parser.add_argument('-e', '--email', type=str, default=None,
                         help="<option> recipient email address to send error messages")                        
-    parser.add_argument('--check', dest='check', default='store_true',
+    parser.add_argument('--check', dest='check', action='store_true',
                         help="Check processed files to see if the files have changed since last processing them")
-    parser.add_argument('--no-check', dest='check', action='store_false',
-                        help="Do not check processed files to see if the files have changed since last processing them")
-    parser.set_defaults(check=False)
+    parser.add_argument('--replace', dest='replace', action='store_false',
+                        help="Remove previous result files")
+    parser.add_argument('--rerun', dest='rerun', action='store_true',
+                        help="Process result files again (generate JSON, csv and barplots again)")
+    parser.add_argument('--runs', nargs="*", default=[],
+                        help="Runs to process again")
+    parser.set_defaults(check=False, replace=True, rerun=False)
 
     return parser.parse_args()
 
@@ -366,13 +399,20 @@ if __name__ == '__main__':
     except:
         cb.callback("Could not update submodules", level='ERROR')
 
-    cursor, connection = open_connection(args.db, callback=cb.callback)
-
     files = glob.glob("{}/**/*_R1_*.fastq.gz".format(args.indir), recursive=True)
-    new_files, runs = get_files(cursor, files, args.ignore_list, args.check, callback=cb.callback)
-    process_files(cursor, args.indir, args.outdir, new_files, binpath=args.binpath, cutabin=args.cutabin,
-                  callback=cb.callback)
-    run_scripts(runs, args.indir, args.outdir, callback=cb.callback)
+
+    if args.rerun:
+        runs = get_runs(files, args.ignore_list, args.runs, callback=cb.callback)
+    else:
+        cursor, connection = open_connection(args.db, callback=cb.callback)
+        new_files, runs = get_files(cursor, files, args.ignore_list, args.check, callback=cb.callback)
+        process_files(cursor, args.indir, args.outdir, new_files, binpath=args.binpath, cutabin=args.cutabin,
+                      callback=cb.callback)
+        connection.commit()
+        connection.close()  
+
+    # Generate result files
+    run_scripts(runs, args.indir, args.outdir, args.replace, callback=cb.callback)
 
     if len(error_msgs) > 0 and args.email is not None:
         cb.callback("Sending error message")
@@ -399,7 +439,4 @@ if __name__ == '__main__':
     if len(new_files) == 0:
         cb.callback("No new data files")
     else:
-        cb.callback("All Done!")
-    
-    connection.commit()
-    connection.close()
+        cb.callback("All Done!")
